Refactor : login 관련 요청은 모두  /auth 라우터로 분리
page 관련 router는 indexRouter
인증 관련 router는 authRouter
