+++ views/signin.pug
@@ -2,7 +2,7 @@ extends layout
 
 block content
   h1.page= page
-  form.form-signin(action='/login' method='POST')
+  form.form-signin(action='/auth/local' method='POST')
     .form-group
       label(for='id') ID
       input#id(type='text' name='id' placeholder='id를 입력하세요' required='' autofocus='')
@@ -11,4 +11,4 @@ block content
       input#password(type='password' name='password' placeholder='Password를 입력하세요' required='')
     hr
     button(type='submit') LOGIN
-  button#btn-register(onclick="location.href='/register'") SIGN IN
+  button#btn-register(onclick="location.href='/signup'") SIGN UP
